backend/requirements.txt
Django==5.0.6
channels==4.1.0
channels-redis==4.2.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
google-generativeai==0.7.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.8
pytest==8.3.2
pytest-django==4.8.0
pytest-asyncio==0.23.8
factory-boy==3.3.0
drf-spectacular==0.27.2
